+++ src/logging/transports/BaseTransport.ts
@@ -1,4 +1,4 @@
-import type { LogEntry } from './LogEntry.js';
+import type { LogEntry } from '../LogEntry.js';
 
 export interface BaseTransport {
   log(logEntry: LogEntry): void;
